updated: profile metadata from contract
