+++ client/app/layout.jsx
@@ -14,8 +14,14 @@ const geistMono = Geist_Mono({
 });
 
 export const metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app"
+  title: {
+    template: "%s | LinkFolio",
+    default: "LinkFolio",
+    absolute: "LinkFolio"
+  },
+  description:
+    "Create and own your digital identity as a soulbound NFT with on-chain metadata, gas-free via NERO Chain’s Paymaster and account abstraction.",
+  keywords: "LinkFolio, profiles, social media, blockchain, portfolio"
 };
 
 export default function RootLayout({ children }) {
