+++ client/app/utils/index.js
@@ -34,7 +34,8 @@ const linkFolioContractABI = [
   "function deleteProfile(uint256 _tokenId)",
   "function getProfileByHandle(string _handle) view returns (uint256 tokenId, string name, string handle, string bio, string avatar, address owner, string[] linkKeys, string[] links)",
   "function handleToTokenId(string handle) view returns (uint256 tokenId)",
-  "function profileExists(string handle) view returns (bool isExists)"
+  "function profileExists(string handle) view returns (bool isExists)",
+  "function tokenURI(uint256 tokenId) view returns (string uri)"
 ];
 
 // nero testnet provider
